FixMyStreet.po: Add pointer to specify address
After having discussed this issue, we decided not to add a separate
address field for the time-being. We now encourage users in the "tip"
in the description field. (The alternative was to do it in the
introductory paragraph of the problem form, but that made the form
seem too cluttered.)

Fixes #9.